Standardize gempak symlinks

+++ gempak/ush/gdas_meta_na.sh
@@ -12,7 +12,9 @@ device="nc | gdas.meta"
 # TODO: Replace this
 #
 export COMIN="${RUN}.${PDY}${cyc}"
-ln -sf "${COM_ATMOS_GEMPAK_1p00}" "${COMIN}"
+if [[ ! -L "${COMIN}" ]]; then
+    ln -sf "${COM_ATMOS_GEMPAK_1p00}" "${COMIN}"
+fi
 
 if [[ "${envir}" == "para" ]] ; then
    export m_title="GDASP"
