+++ gempak/ush/gfs_meta_comp.sh
@@ -26,11 +26,18 @@ export COMIN="gfs.multi"
 mkdir "${COMIN}"
 for cycle in $(seq -f "%02g" -s ' ' 0  "${STEP_GFS}" "${cyc}"); do
     YMD=${PDY} HH=${cycle} GRID="1p00" generate_com gempak_dir:COM_ATMOS_GEMPAK_TMPL
-    ln -s "${gempak_dir}/gfs_1p00_${PDY}${cycle}f"* "${COMIN}"
+    for file_in in "${gempak_dir}/gfs_1p00_${PDY}${cycle}f"*; do
+        file_out="${COMIN}/$(basename "${file_in}")"
+        if [[ -L "${file_out}" ]]; then
+            ln -sf "${file_in}" "${file_out}"
+        fi
+    done
 done
 
 export HPCNAM="nam.${PDY}"
-ln -sf "${COMINnam}/nam.${PDY}/gempak" "${HPCNAM}"
+if [[ ! -L ${HPCNAM} ]]; then
+    ln -sf "${COMINnam}/nam.${PDY}/gempak" "${HPCNAM}"
+fi
 
 #
 # DEFINE YESTERDAY
@@ -91,9 +98,10 @@ for gareas in US NP; do
 
         # Create symlink in DATA to sidestep gempak path limits
         HPCGFS="${RUN}.${init_time}"
-        if [[ -L ${HPCGFS} ]]; then rm "${HPCGFS}"; fi
-        YMD="${init_PDY}" HH="${init_cyc}" GRID="1p00" generate_com source_dir:COM_ATMOS_GEMPAK_TMPL
-        ln -sf "${source_dir}" "${HPCGFS}"
+        if [[ ! -L ${HPCGFS} ]]; then
+            YMD="${init_PDY}" HH="${init_cyc}" GRID="1p00" generate_com source_dir:COM_ATMOS_GEMPAK_TMPL
+            ln -sf "${source_dir}" "${HPCGFS}"
+        fi
 
         if [[ ${init_PDY} == "${PDY}" ]]; then
             desc="T"
@@ -220,9 +228,10 @@ EOF
         ukmet_date=$(date --utc +%Y%m%d%H -d "${PDY} ${cyc} - 12 hours")
         ukmet_PDY=${ukmet_date:0:8}
         ukmet_cyc=${ukmet_date:8:2}
-        if [[ -L ukmet ]]; then rm "ukmet"; fi
-        ln -sf "${COMINukmet}/ukmet.${ukmet_PDY}/gempak" ukmet
-        export HPCUKMET=ukmet
+        export HPCUKMET=ukmet.${ukmet_PDY}
+        if [[ ! -L "${HPCUKMET}" ]]; then
+            ln -sf "${COMINukmet}/ukmet.${ukmet_PDY}/gempak" "${HPCUKMET}"
+        fi
         grid2="F-UKMETHPC | ${ukmet_PDY:2}/${ukmet_date}"
 
         for fhr in 0 12 24 84 108; do
