+++ gempak/ush/gfs_meta_hur.sh
@@ -22,7 +22,9 @@ device="nc | ${metaname}"
 # TODO: Replace this
 #
 export COMIN="${RUN}.${PDY}${cyc}"
-ln -sf "${COM_ATMOS_GEMPAK_1p00}" "${COMIN}"
+if [[ ! -L ${COMIN} ]]; then
+    ln -sf "${COM_ATMOS_GEMPAK_1p00}" "${COMIN}"
+fi
 
 #
 # DEFINE YESTERDAY
@@ -301,13 +303,20 @@ EOF
 export err=$?;err_chk
 
 if [[ ${cyc} -eq 00 ]] ; then
-    ln -sf "${COMINecmwf}/ecmwf.${PDY}/gempak" "ecmwf.${PDY}"
-    ln -sf "${COMINecmwf}/ecmwf.${PDYm1}/gempak" "ecmwf.${PDYm1}"
-    ln -sf "${COMINukmet}/ukmet.${PDY}/gempak" "ukmet.${PDY}"
-    export HPCECMWF=ecmwf
-    export HPCUKMET=ukmet
+    export HPCECMWF=ecmwf.${PDY}
+    HPCECMWF_m1=ecmwf.${PDY}
+    export HPCUKMET=ukmet.${PDYm1}
+    if [[ ! -L "${HPCECMWF}" ]]; then
+        ln -sf "${COMINecmwf}ecmwf.${PDY}/gempak" "${HPCECMWF}"
+    fi
+    if [[ ! -L "${HPCECMWF_m1}" ]]; then
+        Ln -sf "${COMINecmwf}ecmwf.${PDYm1}/gempak" "${HPCECMWF_m1}"
+    fi
+    if [[ ! -L "${HPCUKMET}" ]]; then
+        ln -sf "${COMINukmet}/ukmet.${PDYm1}/gempak" "${HPCUKMET}"
+    fi
     grid1="F-${MDL} | ${PDY:2}/${cyc}00"
-    grid2="ecmwf.${PDYm1}/ecmwf_glob_${PDYm1}12"
+    grid2="${HPCECMWF_m1}/ecmwf_glob_${PDYm1}12"
     grid3="F-UKMETHPC | ${PDY:2}/${cyc}00"
     for fhr in $(seq -s ' ' 12 24 108); do
         gfsfhr=F$(printf "%02g" "${fhr}")
