+++ gempak/ush/gfs_meta_nhsh.sh
@@ -14,7 +14,9 @@ cp "${HOMEgfs}/gempak/fix/datatype.tbl" datatype.tbl
 # TODO: Replace this
 #
 export COMIN="${RUN}.${PDY}${cyc}"
-ln -sf "${COM_ATMOS_GEMPAK_1p00}" "${COMIN}"
+if [[ ! -L ${COMIN} ]]; then
+    ln -sf "${COM_ATMOS_GEMPAK_1p00}" "${COMIN}"
+fi
 
 if [[ "${envir}" == "para" ]] ; then
    export m_title="GFSP"
